test(#1296): pin frozen-list insertion-order (no-sort) contract + perf note
Close 2 LOW roborev items on the recursive collection-order comparator:

- LOW 1 (coverage): the most error-prone part of the contract — frozen
  lists PRESERVE insertion order while sets/maps canonicalize (sort) — was
  untested. Add unit tests (frozen_list_preserves_insertion_order_not_sorted,
  frozen_list_compared_elementwise_positionally) proving compare_list does NOT
  sort ([3,1] vs [1,3] -> Greater, distinct from the same elements as a SET
  which compare Equal) and compares element-wise positionally. Add an e2e
  serialize test (serialize_frozen_list_preserves_insertion_order_not_sorted)
  proving the write path emits a frozen<list<int>> [3,1,2] in insertion order,
  not sorted. Verified all three fail when compare_list / the list-serialize
  arm is made to sort.
- LOW 2 (perf ack, comment-only): document in compare_set/compare_map that each
  pairwise comparison re-sorts both inner collections O(N log N) times under an
  outer sort — correct but redundant; future canonicalize-once optimization
  noted, no behavioral change.

Move the #1296 e2e serialize tests into a dedicated tests/collection_order_serialize.rs
so scenarios_3.rs stays under the campsite test-size threshold (#1135).
