+++ cqlite-core/src/storage/sstable/writer/data_writer/tests/mod.rs
@@ -1,6 +1,7 @@
 //! Relocated `data_writer` unit tests (issue #1118).
 //! Shared fixtures live in `support`; tests are grouped by file.
 
+mod collection_order_serialize;
 mod scenarios_1;
 mod scenarios_2;
 mod scenarios_3;
